Lab 4: update lab04.py with final code; remove stray main.py
